Had to rush to meet deadline
This is going to be a long explanation into one push but here I go. I Fixed spacing because I know I got points taken off last semester. FINALLY fixed the desktop view, it would only accept if I put min-width on top instead of on bottom of max-width. I added a paragraph line and change the placeholder for date to all caps. I added everything to become centered as well as to keep the same font and made the title and the paragraph line have a bit of spaceing from the rest of the form using padding-top and padding-bottom. To top it off I used a Roboto font which is originaly from google. I had the background color gray and the text letters a dark blue. If everything checks with no errors out this would be my last push.